acf-bitcoin/requirements.txt
pandas==2.1.4
numpy==1.26.2
statsmodels==0.14.1
matplotlib==3.8.2
python-binance==1.0.19
scikit-learn==1.3.2 